backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
aiofiles
python-dotenv
jinja2
pillow
